Edits to FLA Files (finally)
Finally figured out that you have to "Break Apart" the items in the timeline to actually see their pieces.  Not a fan of the "IDE", would much rather write UI code.

Regardless, updates to all screens are now complete.  Code is merged (hopefully correctly) and all that remains is to test as extensively as possible... with user support!  :)